Add a QuickCheck unlex-lex test.
This is the first of a small series of QuickCheck tests for the
lexer. The basic idea of these tests is to generate a list of tokens,
convert that list to input text, and then use the lexer to read the
tokens back; we should get the same list of tokens we started out
with.

In this patch, I've added two test modules

 TokenGen

  which contains the generators for the Token type, as well as a
  simple 'unlex' function that converts a Token to text

 LexerQCTest

  which contains the QuickCheck tests themselves

There are actually several possible unlex-lex tests that would vary
how they generate text from tokens, and how they separate tokens when
converting lists of them to text. In this first test, I separate
tokens by single spaces, and I don't bother varying how number and
string tokens are unlexed. For example, I don't randomly generate hex
or octal literals for integer tokens.

This initial test doesn't yet pass. I'll address a couple of lexer
issues the tests have uncovered in the next patch. Since the tests
don't pass, I haven't hooked them into the main LexerTest module.