Don't generate characters outside the ASCII range in character and string literals.
It looks like Alex lexers are restricted to charcters in the range 0x00 to 0xff,
that is, the characters of the Latin1 block.

For the time being, I've restricted to the ASCII range the characters
generated by the 'arbitrary' token generator. I'll likely expand that
to the full Latin1 range once the tests pass without error.