Fix a bug in the lexer's string literal regex.
The rule for string literals had a bug in which it would fail to
correctly lex this text:

 "\\" ""

that is, a string containing the escape for the backslash (\)
character, followed by another string. Any other string would do, but
the empty one is the shortest example.

A couple of patches back fixed the '@charEscape' macro to permit the
escape \\. However, there was a bug in the regex that matched string
literals. The regex was

 \" (@charEscape | ~[\"])* \"

The second alternative for the contents

 ~[\"]

excludes only the quotation mark. It should be

 ~[\\\"]

so that it also excludes the backslash. With the original regex, the
input text

 "\\" ""

would lex as a single string followed by an extra quotation mark. The
string's contents are

 a backslash (\)
 a quotation mark (\")
 a space ( )

The problem is that a lone backslash is permitted by ~[\"].

With the new regex

 \" (@charEscape | ~[\\\"])* \"

backslashes can appear only on character escapes, as they should.

In the process, also I merged the two character regexes into one

 ' (@charEscape | ~[\\\']) '

I've also added

 "\\" ""

to the HUnit lexer tests.