Rewrite the lexer's 'convChar' function.
The 'convChar' utility function

 convChar :: String -> String

used to convert a quoted sequence of Java character literals into a
sequence of Haskell character literals that could be read by 'read'
for either Char or String. The QuickCheck unlex-lex test revealed that
this conversion had a couple of bugs, including forgetting support for
octal escapes, and the accidental capture of decimal digits that trail
a unicode escape.

Since the correct conversion to 'read'able strings is in itself
slightly tricky, I've changed 'convChar' so it converts an unquoted
sequence of Java literals directly to Chars without going through
'read'. This makes the function longer, but a bit more direct and
easier to debug.

At this point, I've run the first unlex-lex test for 100000 iterations
without finding any errors. There are some things that I'd like to
tweak in the implementation of 'convChar' though:

- 'dropQuotes' may be a waste of space and time for long strings;
  'convChar' should deal with the leading and trailing quotation marks itself

- although 'lexicalError' should never be called if 'convChar' is
  consistent with the lexer rules for Java character and string
  literals, it would be better to pass in the position and update the
  column so we can include the line and column in the error message

- as noted in the comments, 'convChar' is the wrong place to be
  dealing with unicode escapes.

To deal with Unicode escapes, and indeed with the fully-permitted
range of Unicode characters, I believe we'll need to abandon Alex and
either write the lexer by hand or using Parsec combinators. If I've
understood the Alex user documentation correctly, it permits
characters only in the range 0x00 to 0xff, essentially the Latin1
block. A Parsec lexer may be too slow, so we may be forced to write
the lexer by hand at some point.