Fix a bug in the lexer's multi-line comment rule.
I added two more simple unlex-lex tests, one which separates tokens by
an empty 'traditional' comment bounded by newlines

 \n/**/\n

and another which separates them by a three-line traditional comment
where a star '*' sits on a line by itself

 \n/*\n *\n */\n

or

 /*
  *
  */

The newlines aid reading the unparse and also prevent accidental
one-line comments when unlexing a token sequence that contains
Op_Slash.

This test found a bug in the lexer rule for traditional multi-line
comments. The unlex-lex test would fail for any three (or more) token
sequence, dropping the tokens between the first and last comment. For
example, the generated token list

 [Op_SlashE,LongTok 3,KW_Package]

unlexes to

 /=/**/3L/**/package

but the lexer produces just the Op_SlashE and the KW_Package tokens on
this text.

The original lexer macros were

 @lineterm = [\n\r] | \r\n

 @any = . | @lineterm
 @tradcomm = "/*" @any* "*/"

The problem with this is the rule can capture intermediate closing
delimiters "*/" since '@any' permits both '*' and '/'. In a source
file with more than one traditional comment, the rule will find the
longest possible match, and so the lexer thinks there is one comment
that starts with the first "/*" and ends with the last "*/".

The solution to this is to change the middle of the '@tradcomm' macro
so it permits a '*' only if it is not followed by a '/'. Now the macro
is

 @tradcomm = "/*" ( ~[\*] | \* (~[\/] | \n) | \n )* "*/"

The original '@any' macro is no longer used, so I've removed it. I've
also used '\n' in preference to '@lineterm' here since set complement
'~ s' is defined as

 . # s == ([\x00-\xff] # \n) # s

and so excludes only the '\n' character. This makes the alternatives
within @tradcomm mutually-exclusive, which isn't necessary but (I
think) simplifies reasoning about what the rule matches.

I've run the the unlex-lex tests with intervening spaces and the two
styles of intervening comments for 100000 iterations without
failing. I've included the new unlex-lex tests in this patch.