Add a token for '@', which decorates annotation names.
Chapter 3 of the Java Language Specification does not mention '@', but
sections 9.6 (Annotation types) and 9.7 (Annotations) say that '@' is
a separate token. These sections refer to the character as the
"at-sign".

Since it has the flavor of a prefix operator, I've called the new
token Op_AtSign, and grouped it with the other Op_ tokens.

With these changes, we ought to be able to successfully lex Java
source files that include annotation type declarations and
annotations. The AST types and the parser do not yet support
annotations though.