Fix lexer action for hexadecimal long literals.
The lexer action for hexadecimal long literals of the form

 0[xX] hexdigit+ [lL]

was building an IntTok instead of a LongTok. I've fixed this, and
modified the HUnit tests so that for an Integer value they test

 - lexing IntToks using decimal, octal, and hexadecimal
   representations

 - lexing LongToks using decimal, octal, and hexadecimal
   representations

These tests demonstrated the original bug. Now the bug is gone.

The QuickCheck tests didn't find this bug because they don't generate
input text with octal and hexadecimal integer literals. I found this
by doing some tests using real Java source files that included
hexadecimal long literals (the test code will be forthcoming in a
future patch).

I'll have to modify the QuickCheck tests to allow varying the input
representation for numbers.