Fix the lexer rule for block comments, yet again.
The 'fixed' lexer rule for traditional comments is incorrect, since it
fails for any comment that ends in consecutive stars. For example, in
the input text

 /***/ 1 /* */

the lexer will find no tokens. The problem is that the lexer rule for
traditional comments matches the text like this

 (/*)     begin comment
 (**)     a star followed by a non-slash
 (/ 1 /)  non-stars
 (* )     a star followed by a non-slash
 (*/)     end comment

I found this problem attempting to lex some real source files, but
I've added a few more HUnit test cases to catch the problem.

The fix is to change the lexer rule so it eagerly consumes stretches
of consecutive starts in the middle and at the end. Now the rule is

 "/*" ( ~[\*] | \*+ (~[\/\*] | \n) | \n )* \*+ "/" 
                ^^^      ^^                ^^^^^^^

where (^) marks the changes. Now the middle will not accidentally
match the "**" in "**/", and the new HUnit tests pass.