Make the 'Test' program work again.
Somewhere along the line, a long time ago it looks like, the
'test/Test.hs' program was broken by an unfortunate whitespace error,
and by the removal of 'parseCompilationUnit' from the export list in
'Java.Lang.Parser'.

I've corrected the indentation error, and added 'parseCompilationUnit'
to the exports. Now 'Test' builds and superficially seems to do what
it's supposed to on the "*.java" files in the 'test' directory.