Add 'GHC-options', 'GHC-prof-options' in the cabal file.
We might as well turn on optimization and the standard batch of
warnings.  Unfortunately, the alex-generated lexer has a lot of
warnings, but it's undoubtedly useful for the hand-written
stuff. There are some warnings in the 'Parser' and 'Pretty' modules
that look like they warrant attention, but I'll wait to fix those.

I've also added 'GHC-prof-options' with '-auto-all' and '-caf-all',
since I've found those useful in the past.