Fix for parsing error when calling method for new instance.
Parsing of this:

class Something {
   void a() {
      new Object().a();
   }
}

fails, however, this is correct java code.

The problem is that instance creation parser tried first, which causes
instance creation be parsed, instead of method invocation of new
instance. This makes '.a()' left unexpected.

To fix this, order of parsings is changed - method invocation tried first.