Fix for multidimension array creation.
There was a problem when parsing constructions like

Object[][] a = new Object[1][] ;

The problem was because second square brackets in

new Object[1][]

was tried by brackets function, which wasn't wrapped in 'try'. This caused
it consume '[', and then fail. So parsing of next '[]' failed, because
expecting '[', but getting ']' (as '[' already consumed).